+++ models/tt_transformers/conftest.py
@@ -1,0 +1,19 @@
+# SPDX-FileCopyrightText: © 2025 Tenstorrent Inc.
+
+# SPDX-License-Identifier: Apache-2.0
+import pytest
+
+import ttnn
+
+
+@pytest.fixture
+def device_params(request, galaxy_type):
+    # Get param dict passed in from test parametrize (or default to empty dict)
+    params = getattr(request, "param", {}).copy()
+
+    if "fabric_config" in params and params["fabric_config"] == True:
+        params["fabric_config"] = (
+            ttnn.FabricConfig.FABRIC_1D_RING if galaxy_type == "6U" else ttnn.FabricConfig.FABRIC_1D
+        )
+
+    return params
